% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_nav_tabs.R
\name{ggplot_to_html}
\alias{ggplot_to_html}
\title{ggplot_to_html}
\usage{
ggplot_to_html(ggplot, width = 7, height = 3.5)
}
\arguments{
\item{ggplot}{}

\item{width}{}

\item{height}{}
}
\description{
ggplot_to_html
}
