% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_document.R
\name{html_document}
\alias{html_document}
\title{R6 Class representing a HTML Report that can modify and render itself}
\description{
R6 Class representing a HTML Report that can modify and render itself

R6 Class representing a HTML Report that can modify and render itself
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-html_document-new}{\code{html_document$new()}}
\item \href{#method-html_document-add_object}{\code{html_document$add_object()}}
\item \href{#method-html_document-add_text}{\code{html_document$add_text()}}
\item \href{#method-html_document-add_break}{\code{html_document$add_break()}}
\item \href{#method-html_document-add_df_info}{\code{html_document$add_df_info()}}
\item \href{#method-html_document-open}{\code{html_document$open()}}
\item \href{#method-html_document-reload}{\code{html_document$reload()}}
\item \href{#method-html_document-render}{\code{html_document$render()}}
\item \href{#method-html_document-save}{\code{html_document$save()}}
\item \href{#method-html_document-help}{\code{html_document$help()}}
\item \href{#method-html_document-clone}{\code{html_document$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-html_document-new"></a>}}
\if{latex}{\out{\hypertarget{method-html_document-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{html_document$new(
  title = "",
  author = "",
  theme = "default",
  bs_version = 5,
  filemanager = "browseURL"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-html_document-add_object"></a>}}
\if{latex}{\out{\hypertarget{method-html_document-add_object}{}}}
\subsection{Method \code{add_object()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{html_document$add_object(
  object,
  name = NULL,
  header = NULL,
  footer = NULL,
  width = 8,
  height = 4,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-html_document-add_text"></a>}}
\if{latex}{\out{\hypertarget{method-html_document-add_text}{}}}
\subsection{Method \code{add_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{html_document$add_text(text, level = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-html_document-add_break"></a>}}
\if{latex}{\out{\hypertarget{method-html_document-add_break}{}}}
\subsection{Method \code{add_break()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{html_document$add_break()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-html_document-add_df_info"></a>}}
\if{latex}{\out{\hypertarget{method-html_document-add_df_info}{}}}
\subsection{Method \code{add_df_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{html_document$add_df_info(data, id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-html_document-open"></a>}}
\if{latex}{\out{\hypertarget{method-html_document-open}{}}}
\subsection{Method \code{open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{html_document$open()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-html_document-reload"></a>}}
\if{latex}{\out{\hypertarget{method-html_document-reload}{}}}
\subsection{Method \code{reload()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{html_document$reload()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-html_document-render"></a>}}
\if{latex}{\out{\hypertarget{method-html_document-render}{}}}
\subsection{Method \code{render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{html_document$render()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-html_document-save"></a>}}
\if{latex}{\out{\hypertarget{method-html_document-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{html_document$save(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-html_document-help"></a>}}
\if{latex}{\out{\hypertarget{method-html_document-help}{}}}
\subsection{Method \code{help()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{html_document$help()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-html_document-clone"></a>}}
\if{latex}{\out{\hypertarget{method-html_document-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{html_document$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
