% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_document.R
\name{pptx_document}
\alias{pptx_document}
\title{R6 Class representing a power point presentation that can modify itself}
\description{
R6 Class representing a power point presentation that can modify itself

R6 Class representing a power point presentation that can modify itself
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pptx_document-new}{\code{pptx_document$new()}}
\item \href{#method-pptx_document-add_ggplot}{\code{pptx_document$add_ggplot()}}
\item \href{#method-pptx_document-add_flextable}{\code{pptx_document$add_flextable()}}
\item \href{#method-pptx_document-open}{\code{pptx_document$open()}}
\item \href{#method-pptx_document-reload}{\code{pptx_document$reload()}}
\item \href{#method-pptx_document-save}{\code{pptx_document$save()}}
\item \href{#method-pptx_document-clone}{\code{pptx_document$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pptx_document-new"></a>}}
\if{latex}{\out{\hypertarget{method-pptx_document-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pptx_document$new(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pptx_document-add_ggplot"></a>}}
\if{latex}{\out{\hypertarget{method-pptx_document-add_ggplot}{}}}
\subsection{Method \code{add_ggplot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pptx_document$add_ggplot(
  gg,
  header = "Titel",
  align = "center",
  current_slide = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pptx_document-add_flextable"></a>}}
\if{latex}{\out{\hypertarget{method-pptx_document-add_flextable}{}}}
\subsection{Method \code{add_flextable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pptx_document$add_flextable(
  tabelle,
  font_size = 30,
  header = "Titel",
  align = "center",
  current_slide = F,
  caption = NULL,
  footer = NULL,
  digits = 2,
  percent_vector = NULL,
  theme_fun = theme_booktabs
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pptx_document-open"></a>}}
\if{latex}{\out{\hypertarget{method-pptx_document-open}{}}}
\subsection{Method \code{open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pptx_document$open()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pptx_document-reload"></a>}}
\if{latex}{\out{\hypertarget{method-pptx_document-reload}{}}}
\subsection{Method \code{reload()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pptx_document$reload()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pptx_document-save"></a>}}
\if{latex}{\out{\hypertarget{method-pptx_document-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pptx_document$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pptx_document-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pptx_document-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pptx_document$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
