% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.R
\name{presentation}
\alias{presentation}
\title{R6 Class representing a power point presentation that can modify itself}
\description{
R6 Class representing a power point presentation that can modify itself

R6 Class representing a power point presentation that can modify itself
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{presentation$new()}}
\item \href{#method-add_ggplot}{\code{presentation$add_ggplot()}}
\item \href{#method-add_flextable}{\code{presentation$add_flextable()}}
\item \href{#method-open}{\code{presentation$open()}}
\item \href{#method-reload}{\code{presentation$reload()}}
\item \href{#method-save}{\code{presentation$save()}}
\item \href{#method-clone}{\code{presentation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{presentation$new(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_ggplot"></a>}}
\if{latex}{\out{\hypertarget{method-add_ggplot}{}}}
\subsection{Method \code{add_ggplot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{presentation$add_ggplot(
  gg,
  header = "Titel",
  align = "center",
  current_slide = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_flextable"></a>}}
\if{latex}{\out{\hypertarget{method-add_flextable}{}}}
\subsection{Method \code{add_flextable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{presentation$add_flextable(
  tabelle,
  font_size = 30,
  header = "Titel",
  align = "center",
  current_slide = F,
  caption = NULL,
  footer = NULL,
  digits = 2,
  percent_vector = NULL,
  theme_fun = theme_booktabs
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open"></a>}}
\if{latex}{\out{\hypertarget{method-open}{}}}
\subsection{Method \code{open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{presentation$open()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reload"></a>}}
\if{latex}{\out{\hypertarget{method-reload}{}}}
\subsection{Method \code{reload()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{presentation$reload()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{presentation$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{presentation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
