% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_table.R
\name{quick_table}
\alias{quick_table}
\title{quickly generate a nicely formatted flextable object}
\usage{
quick_table(
  tab,
  fontsize = 11,
  caption = NULL,
  footer = NULL,
  digits = 2,
  percent_vector = NULL,
  as_reporter_table = F,
  theme_fun = function(x) {     theme_booktabs(x, bold_header = T) },
  total_width = 6.267717
)
}
\arguments{
\item{tab}{A dataframe/tibble containing the table information}

\item{fontsize}{Fontsize used for the entire table}

\item{caption}{An optional caption}

\item{footer}{An optional footer}

\item{digits}{The number of digits to round to}

\item{percent_vector}{A vector (e.g. c(0.1,0.2,0.8)) that sums to 1 and identicates how wide each column should be relative to the others}

\item{as_reporter_table}{Return as report table R6 instance for further processing. By default returns a flextable object}

\item{theme_fun}{A flextable theme function, by defaut booktab with bold header}

\item{total_width}{Width of the table in inches. Defaults to regular Word document with 1 inch margins}
}
\value{
a flextable object or a reporter_table object if as_reporter_table=T
}
\description{
quickly generate a nicely formatted flextable object
}
\examples{

library(prepkit)
tab1<-prep_freq_table(diamonds,cut)
quick_table(tab1,percent_vector = c(0.5,0.25,0.25),fontsize = 22)

}
