% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_document.R
\name{report_html}
\alias{report_html}
\title{Create a Report Object that can interactively create a rmd html document}
\usage{
report_html(
  title = "",
  author = "",
  theme = "default",
  bs_version = 5,
  filemanager = "browseURL"
)
}
\arguments{
\item{title}{}

\item{author}{}

\item{theme}{a bootswatch theme}

\item{bs_version}{}

\item{filemanager}{}
}
\description{
Create a Report Object that can interactively create a rmd html document
}
