% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_nav_tabs.R
\name{report_nav_bar}
\alias{report_nav_bar}
\title{report_nav_bar}
\usage{
report_nav_bar(
  titles,
  bodies,
  nav_title = "Navigation",
  full_screen = TRUE,
  height = 4.6,
  width = NULL,
  gg_height = 5,
  gg_width = 8
)
}
\arguments{
\item{titles}{navbar titles}

\item{bodies}{list of objects}

\item{height}{height of the container}
}
\description{
report_nav_bar
}
