% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_nav_tabs.R
\name{report_nav_tabs}
\alias{report_nav_tabs}
\title{prep_nav_tabs}
\usage{
report_nav_tabs(titles, bodies, footers = NULL, height = 4.6, width = NULL)
}
\arguments{
\item{titles}{navbar titles}

\item{bodies}{list of objects}

\item{footers}{optional footers}

\item{height}{height of the container}
}
\description{
prep_nav_tabs
}
