% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_reactable.R
\name{report_reactable}
\alias{report_reactable}
\title{report_reactable}
\usage{
report_reactable(
  table,
  fontsize = 14,
  headersize = 16,
  resizable = FALSE,
  groupBy = NULL,
  total_width = 8,
  percent_vector = NULL,
  defaultPageSize = 10,
  ...
)
}
\arguments{
\item{table}{}

\item{fontsize}{}

\item{headersize}{}

\item{resizable}{}

\item{groupBy}{}

\item{total_width}{}

\item{percent_vector}{}

\item{defaultPageSize}{}
}
\description{
report_reactable
}
