% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_document.R
\name{report_word}
\alias{report_word}
\title{report_word}
\usage{
report_word(path)
}
\arguments{
\item{path}{either a filepath or name / name.pptx to create a file based on a template in the current workding directory.
Alternatively you may provide a filepath of an existing pptx to read in the existing file.}
}
\value{
an R6 Class called presentation that can modify itself interactively
}
\description{
report_word
}
