% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter_table.R
\name{reporter_table}
\alias{reporter_table}
\title{R6 Class representin´g a flextable that can modify itself}
\description{
R6 Class representin´g a flextable that can modify itself

R6 Class representin´g a flextable that can modify itself
}
\examples{
library(prepkit)
tab2<-prep_freq_table(diamonds,cut,color)
rtab2<-reporter_table$new(tab2,percent_vector = c(0.5,0.25,0.25),fontsize = 22)
rtab2$round()
rtab2$merge_col_cells(column = "cut",h_lines = T)
rtab2$add_label_row(colwidths = c(2,3,3),values = c("Variables","Frequency","Proportions"),align = "left")
rtab2$add_caption("Table 1: Table on diamonds")
rtab2$add_footer("Footer: this is a table")


}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tab}}{A dataframe/tibble containing the table information}

\item{\code{modified_tab}}{formatted table (e.g. with rouneded value)}

\item{\code{ft}}{flextable representation of reporter table}

\item{\code{nrow}}{number of rows}

\item{\code{ncols}}{number of columns}

\item{\code{width_vector}}{with of each columns as a fector}

\item{\code{fontsize}}{Fontsize used for the entire table}

\item{\code{theme_fun}}{A flextable theme function, by defaut booktab with bold header}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ncols}}{number of columns}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{reporter_table$new()}}
\item \href{#method-update}{\code{reporter_table$update()}}
\item \href{#method-round}{\code{reporter_table$round()}}
\item \href{#method-help}{\code{reporter_table$help()}}
\item \href{#method-add_label_row}{\code{reporter_table$add_label_row()}}
\item \href{#method-add_caption}{\code{reporter_table$add_caption()}}
\item \href{#method-add_footer}{\code{reporter_table$add_footer()}}
\item \href{#method-merge_col_cells}{\code{reporter_table$merge_col_cells()}}
\item \href{#method-save}{\code{reporter_table$save()}}
\item \href{#method-clone}{\code{reporter_table$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new reporter table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporter_table$new(
  tab,
  percent_vector = NULL,
  fontsize = 11,
  theme_fun = function(x) {     theme_booktabs(x, bold_header = T) },
  total_width = 6.267717
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{percent_vector}}{a}

\item{\code{fontsize}}{a}

\item{\code{theme_fun}}{a}

\item{\code{total_width}}{a}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new reporter table object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporter_table$update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-round"></a>}}
\if{latex}{\out{\hypertarget{method-round}{}}}
\subsection{Method \code{round()}}{
round the table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporter_table$round(digits = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{the n of digits to round to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-help"></a>}}
\if{latex}{\out{\hypertarget{method-help}{}}}
\subsection{Method \code{help()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporter_table$help()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_label_row"></a>}}
\if{latex}{\out{\hypertarget{method-add_label_row}{}}}
\subsection{Method \code{add_label_row()}}{
add a header row that can be merged across cells
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporter_table$add_label_row(colwidths, values, align = "center", bold = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{colwidths}}{width vector}

\item{\code{values}}{a}

\item{\code{align}}{a}

\item{\code{bold}}{a}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_caption"></a>}}
\if{latex}{\out{\hypertarget{method-add_caption}{}}}
\subsection{Method \code{add_caption()}}{
add a caption
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporter_table$add_caption(caption)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{a caption}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_footer"></a>}}
\if{latex}{\out{\hypertarget{method-add_footer}{}}}
\subsection{Method \code{add_footer()}}{
add a footer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporter_table$add_footer(footer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{a footer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-merge_col_cells"></a>}}
\if{latex}{\out{\hypertarget{method-merge_col_cells}{}}}
\subsection{Method \code{merge_col_cells()}}{
merge col cells by a variable in tab (should be sorted by that variable!)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporter_table$merge_col_cells(column, h_lines = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column}}{a col in tab as a string}

\item{\code{h_lines}}{bool whether to insert horizontal lines by factor level}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
save the file as docx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporter_table$save(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path to file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporter_table$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
