% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_document.R
\name{word_document}
\alias{word_document}
\title{R6 Class representing a word document that can modify itself}
\description{
R6 Class representing a word document that can modify itself

R6 Class representing a word document that can modify itself
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{word_document$new()}}
\item \href{#method-add_ft}{\code{word_document$add_ft()}}
\item \href{#method-add_ggplot}{\code{word_document$add_ggplot()}}
\item \href{#method-open}{\code{word_document$open()}}
\item \href{#method-reload}{\code{word_document$reload()}}
\item \href{#method-save}{\code{word_document$save()}}
\item \href{#method-clone}{\code{word_document$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{word_document$new(path = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_ft"></a>}}
\if{latex}{\out{\hypertarget{method-add_ft}{}}}
\subsection{Method \code{add_ft()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{word_document$add_ft(flextab)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_ggplot"></a>}}
\if{latex}{\out{\hypertarget{method-add_ggplot}{}}}
\subsection{Method \code{add_ggplot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{word_document$add_ggplot(gg, default_theme = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open"></a>}}
\if{latex}{\out{\hypertarget{method-open}{}}}
\subsection{Method \code{open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{word_document$open()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reload"></a>}}
\if{latex}{\out{\hypertarget{method-reload}{}}}
\subsection{Method \code{reload()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{word_document$reload()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{word_document$save(path = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{word_document$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
